refactor: task-detail feature を prototype.jsx から分離 (#7)
- features/task-detail/TaskDetailPanel.jsx: DetailPanel を移設しリネーム
  - Markdown エディタ/プレビュー切替 (editing, draft の state) はそのまま保持
- features/task-detail/TaskDetailPanel.test.jsx: 9件の smoke test
  - タイトル/プロジェクト名表示、完了/未完了切替ボタン、編集→保存フロー
  - onUpdate(taskId, draft), onToggleDone + onClose の呼び出し、依存バッジ
- prototype.jsx から DetailPanel 宣言を削除 (約287行削減)
- App の `<DetailPanel>` を `<TaskDetailPanel>` に置換

ボトムシート部分は EventDetailPanel と重複するが、本 Issue では共通化せず
個別に持つ（shared/ui/BottomSheet への昇格は将来検討）。
